+++ types/affinda.d.ts
@@ -2091,6 +2091,8 @@ export declare interface DocumentUpdate {
     isArchived?: boolean;
     /** Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese. */
     language?: string;
+    /** Specify a custom identifier for the document. */
+    identifier?: string;
 }
 
 export declare interface DocumentWarning {