+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Endpoints for add/remove tag for documents
+- Identifier field in DocumentUpdate model
 
 ## [6.9.0] - 2023-05-02
 ### Added